Endpoints Creation
Register/Login

Users must be able to create an account and log in.

View Courses

Users should be able to view all the available courses.

Enroll in Course

Users can enroll in courses.

If it's their first enrollment, a student account must be created.

When enrolling, a request must be sent to the Finance microservice to create an invoice.
Finance Microservice

When a student is created, send a request to the Finance microservice to create a Finance account.

When a course is enrolled in, send a request to the Finance microservice to generate an invoice.

Library Microservice
When a student is created, send a request to the Library microservice to create a Library account.